+++ src/main/java/com/example/demo/Municipio/MunicipioNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.demo.Municipio;
 
 class MunicipioNotFoundException extends RuntimeException {
 
-    MunicipioNotFoundException(Long id) {
+    MunicipioNotFoundException(Integer id) {
       super("No se encontro el Municipio " + id);
     }
   }
