+++ src/main/java/com/example/demo/Municipio/MunicipioRepository.java
@@ -2,5 +2,5 @@ package com.example.demo.Municipio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface MunicipioRepository extends JpaRepository<Municipio, Long> {}
+interface MunicipioRepository extends JpaRepository<Municipio, Integer> {}
