+++ src/main/java/com/example/demo/Provincia/ProvinciaNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.demo.Provincia;
 
 class ProvinciaNotFoundException extends RuntimeException {
 
-    ProvinciaNotFoundException(Long id) {
+    ProvinciaNotFoundException(Integer id) {
       super("No se encontro la provincia " + id);
     }
   }
