+++ src/main/java/com/example/demo/Provincia/ProvinciaRepository.java
@@ -2,4 +2,4 @@ package com.example.demo.Provincia;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface ProvinciaRepository extends JpaRepository<Provincia, Long> {}
+interface ProvinciaRepository extends JpaRepository<Provincia, Integer> {}
